+++ AggroSystem/SightAggroable.cs
@@ -2,6 +2,6 @@ namespace AggroSystem
 {
     public interface SightAggroable : SoundAggroable
     {
-        double SightRadius();
+        float SightRadius();
     }
 }
